feat: backlog cleanup — nav-compose, ImportPipeline, Int→Long, strings, tests, landing page (v0.4.0) (#14)
* backlog 1/6: navigation-compose

MainScreen is now a NavHost with routes decks/quest/import/guide + a separate
'study' route (bottom bar hidden). Tab switches use popUpTo(start){saveState}
+ restoreState + launchSingleTop, so each tab keeps its scroll/input state
across switches (the old when(activeTab) lost it). rememberNavController
restores the back stack across config change + process death.

currentDeck in the ViewModel stays the source of truth: a LaunchedEffect
navigates to/from the study route as it changes, so selectDeck() /
clearActiveDeck() and the study BackHandler all behave exactly as before.

Device-verified: 4 tabs switch, deck→study→system-back→decks, bottom-bar
visibility, cold start 1.8s, no navigation exceptions.

* backlog 2/6: extract ImportPipeline from StudyViewModel

The whole import + staged-merge flow (~450 lines) moves to
domain/ImportPipeline: offline JSON + word:meaning parsing, AI parsing with
chunked-PDF handling + recursive subdivision, and the undoable merge into an
existing deck. ImportState / MergePlan move with it.

StudyViewModel keeps thin delegates (importDeckFromRawText, confirmMerge,
cancelMerge, undoLastImport, resetImportState, importState) and drops from
~1013 → ~540 lines. UI refs updated StudyViewModel.ImportState ->
ImportPipeline.ImportState.

Pure code move — bodies copied verbatim, only field/param plumbing changed
(getApplication -> appContext, _nativeLanguage.value -> nativeLanguage(),
etc). Constructor deps compile-checked. Compiles clean (no warnings), 20 unit
tests green, app launches + navigates all screens with no crash. A live
import wasn't re-driven on device (IME/scroll flakiness on this ColorOS
unit); the equivalent JSON-file path was verified earlier this session on
functionally identical code.

* backlog 3/6: Int → Long primary keys (schema v3)

Deck.id / Flashcard.id / Flashcard.deckId / ChatLog.id / ChatLog.deckId /
ChatLog.flashcardId are Long. SQLite stores every INTEGER column as a 64-bit
value already, so Room's generated schema is byte-identical (v2 and v3 share
identity hash fcf2ec48…) — MIGRATION_2_3 is a no-op that only exists to let
Room accept the version bump. Zero data-migration risk.

Ripples: DAO/repo params, _flippedCardIds Set<Long>, _cardModes Map<Long,_>,
getLearningModeForCard(Long), savedState KEY_ACTIVE_DECK as Long, ImportUndo
id lists Long, dropped now-redundant .toInt() on insert results, DeckMergeTest
card ids .toLong().

Device-verified: install v3 over v2 → app launches (1.7s), Master Pool's 142
cards intact, study loop grades + writes fine, no Room/SQLite/cast errors.

* chore: version 0.4.0

* backlog 4/6: strings.xml scaffold + Guide screen + bottom nav externalised

Adds res/values/strings.xml and moves the Guide (TutorialScreen) copy + the
four bottom-nav labels/descriptions to string resources (TopRoute now carries
@StringRes ids). This is the reference pattern.

The remaining ~250 UI strings (Decks / Quest / Import / Study / dialogs /
tutor sheet) and the dynamic error strings built in StudyViewModel /
ImportPipeline are NOT in this commit — a lot of them are interpolated
(format-string args) or raised from non-composable lambdas needing context
plumbing, so each needs a per-screen device check to be sure nothing throws
an IllegalFormatException at render time. That's a focused follow-up, not
safe to batch unreviewed.

* backlog 5/6: regression tests for the import parsing layer

Instead of Roborazzi screenshot tests (pixel diffs are notoriously
CI-environment-flaky and need a dedicated stable-threshold setup), this adds
stable JVM unit coverage for the code the pipeline extraction touched:

- new domain/ImportParsing object — the pure parsers (JSON fence stripping /
  extraction, full-object + bare-array deck parsing, the offline
  colon/hyphen/equals list parser, the YouTube + URL-echo heuristics) split
  out of ImportPipeline so they're testable without the whole pipeline.
- ImportParsingTest: 10 cases covering the happy paths, the fallbacks, and
  the throw-on-junk path.

Unit tests: 14 → 24, all green. A Roborazzi visual-regression suite with a
tuned comparison threshold remains a separate task.

* backlog 6/6: GitHub Pages landing page (docs/index.html)

Single self-contained static page (no build step, no external deps,
light/dark aware) — feature overview, the bring-your-own-key positioning,
the study loop, an architecture note, and a link to releases/latest.

To publish: repo Settings → Pages → Source: 'Deploy from a branch',
branch main, folder /docs. The CI push filter doesn't include docs/, so
edits here never trigger a build or release.

+++ app/src/main/java/io/github/playfoundryhq/adaptiveflow/ui/QuestScreen.kt
@@ -882,7 +882,7 @@ fun ResponsiveVocabularyGrid(
 ) {
     val c = AppTheme.colors
     // State collection managing individual card flips
-    var flippedCardIds by remember { mutableStateOf<Set<Int>>(emptySet()) }
+    var flippedCardIds by remember { mutableStateOf<Set<Long>>(emptySet()) }
 
     LazyVerticalGrid(
         columns = GridCells.Adaptive(minSize = 130.dp),
